+++ .gitignore
@@ -1,4 +1,3 @@
-lazy-lock.json
 tt.*
 .tests
 doc/tags
