+++ README.md
@@ -7,11 +7,16 @@
 ## ⚠️ Requirements
 
 - Up to Date package 📦
+<!-- TODO: gonna add another package needed ASAP -->
 - [git](https://github.com/git-guides/install-git)
 - [neovim](https://github.com/neovim/neovim/wiki/Installing-Neovim)
-- **clang** compiler
+- **c/cpp** compiler
+  - _clang_
+  - _gcc_
+- make
 - [node.js](https://github.com/nodejs/node)
 - [ripgrep](https://github.com/BurntSushi/ripgrep)
+- [gitui](https://github.com/extrawurst/gitui) -- i changes lazygit to gitui
 
 ## 🚀 Installation
