+++ lua/config/lazy.lua
@@ -1,3 +1,4 @@
+-- stylua: ignore
 local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
 if not vim.loop.fs_stat(lazypath) then
   -- bootstrap lazy.nvim
@@ -33,7 +34,7 @@ require("lazy").setup({
     -- version = "*", -- try installing the latest stable version for plugins that support semver
   },
   install = { colorscheme = { "catppuccin", "habamax" } },
-  checker = { enabled = true }, -- automatically check for plugin updates
+  checker = { enabled = false }, -- disable auto check for plugin updates
   performance = {
     rtp = {
       -- disable some rtp plugins
