+++ lua/plugins/ui.lua
@@ -1,0 +1,8 @@
+return {
+  {
+    "echasnovski/mini.indentscope",
+    opts = {
+      draw = { animation = require("mini.indentscope").gen_animation.none() },
+    },
+  },
+}
